[FIX] SelfTest ViewModel 분리
